### 2016-11-23 #121 @whedon
[PRE REVIEW]: vbvs.concurrent: Fitting Methods for the Functional Linear Concurrent Model
**Submitting author:** @jeff-goldsmith (<a href="http://orcid.org/0000-0002-6150-8997">Jeff Goldsmith</a>)
**Repository:** <a href="https://github.com/jeff-goldsmith/vbvs.concurrent" target ="_blank">https://github.com/jeff-goldsmith/vbvs.concurrent</a>
**Version:** v0.1
