### 2016-11-28 #121 @jeff-goldsmith
Here are some folks who would be able to review the code and have expertise in the statistical area the package targets:
@fabian-s, @huangracer, @jgellar, @swihart, @Jfortin1, @mandymejia, @emsweene, @sbrockhaus.

Full disclosure -- I've worked with several of these people on other projects, especially those who have contributed to the [refund](https://github.com/refunders/refund) package. If nothing else this might give you something to go on. Thanks!
