### 2016-12-13 #121 @cMadan
@fabian-s, @huangracer, @jgellar, @swihart, @Jfortin1, @mandymejia, @emsweene, @sbrockhaus - would any of you be interested in reviewing this package? It's fine if you've collaborated with @jeff-goldsmith, we just need to make sure the code is sufficiently described, documented, and functional (see http://joss.theoj.org/about#reviewer_guidelines)
