### 2016-12-13 #121 @fabian-s
I'd be willing to review it, but won't be able to make it before January. Is that acceptable?
