### 2016-12-15 #141 @cMadan
@fabian-s - The reviewer guidelines are here: http://joss.theoj.org/about#reviewer_guidelines

For the most part you need to see if the statements above in the first post are true and check things off accordingly. If something isn't the case, or you otherwise have a suggestion, either post a comment to @jeff-goldsmith here (or ask me) or make an issue in the project's repository (https://github.com/jeff-goldsmith/vbvs.concurrent) and we'll sort things out until the submission meets JOSS' guidelines.
