### 2016-12-16 #141 @fabian-s
@jeff-goldsmith:

- In paper.md:
    - typo: line 20: "subjct"
    - please add author and affiliation
    - please add DOI for Wrobel, Park, Staicu, Goldsmith (2016)
- the shiny app does not show CIs for the estimated coefficient functions -- please correct the description of the plot given by `plot_shiny`.
