### 2016-12-16 d550ab7
fix typo; add doi for refund.shiny paper in Stat
