+++ paper.bib
@@ -18,5 +18,6 @@
   title = {Interactive graphics for functional data analyses},
   volume = {5},
   pages = {88-101},
-  Year = {2016}
+  year = {2016},
+  doi = {10.1002/sta4.109}
 }
