+++ paper.md
@@ -1,5 +1,6 @@
 ---
 title: 'vbvs.concurrent: Fitting Methods for the Functional Linear Concurrent Model'
+date: "16 December 2016"
 tags:
 - statistical analysis
 - R
@@ -7,7 +8,6 @@ authors:
 - affiliation: Columbia University
   name: Jeff Goldsmith
   orcid: 0000-0002-6150-8997
-date: 17 November 2016
 bibliography: paper.bib
 ---
 
@@ -17,7 +17,7 @@ Functional data analysis is concerned with understanding measurements made over 
 $$
 Y_i(t) = \beta_0(t) + \sum_{k = 1}^{p}X_{ik}(t)\beta_k(t) + \delta_i(t)
 $$
-where $Y_i(t)$ is the functional response for subjct $i$, the $X_{ik}(t)$ are functional predictors, the $\beta_k(t)$ are functional coefficients of interest, and the $\delta_i(t)$ are possibly correlated errors. This package implements two statistical methods (with and without variable selection) for estimating the parameters in the functional linear concurrent model; these methods are described in detail [@goldsmith2017]. 
+where $Y_i(t)$ is the functional response for subject $i$, the $X_{ik}(t)$ are functional predictors, the $\beta_k(t)$ are functional coefficients of interest, and the $\delta_i(t)$ are possibly correlated errors. This package implements two statistical methods (with and without variable selection) for estimating the parameters in the functional linear concurrent model; these methods are described in detail [@goldsmith2017]. 
 
 Given tidy datasets containing functional responses and predictors for all subjects, `vb_concurrent` and `vbvs_concurrent` fit the functional linear concurrent model using variational Bayes and variational Bayes with variable selection, respectively. These functions produce objects of class `flcm` and have the same structure. Coefficients and predictions can be extracted or computed from `flmc` objects using `coef` and `predict`. Interactive visualizations of `flmc` objects are supported through the `refund.shiny` package [@refund.shiny, @wrobel2016]. 
 
