### 2016-12-16 #141 @jeff-goldsmith
@fabian-s: thanks for the quick review!

> - In paper.md:
>   - typo: line 20: "subjct"
>   - please add author and affiliation
>   - please add DOI for Wrobel, Park, Staicu, Goldsmith (2016)

the first and third of these are addressed in jeff-goldsmith/vbvs.concurrent@d550ab78ab1aa0f8ca4fea1f9951984bbd3f33ab.

the second seems to be an issue with the default latex template in rstudio, which may not support the header requested by JOSS; perhaps @cMadan can comment?

> - the shiny app does not show CIs for the estimated coefficient functions -- please correct the description of the plot given by `plot_shiny`.

great catch! this is addressed in refunders/refund.shiny@e147d70c88fc4f4f061bbc77d891ce31562021c8.
