### 2016-12-19 #141 @cMadan
@jeff-goldsmith - I looked at the paper.md (https://raw.githubusercontent.com/jeff-goldsmith/vbvs.concurrent/master/paper.md) and I think you're okay. A few other recent accepted submissions similarly didn't have separate author/affiliation sections in the header, so it should compile fine.
