### 2016-12-19 #141 @cMadan
@fabian-s - it looks like the submission is all set?

@jeff-goldsmith - if everything is all set, the last thing we'll need is for the repo to be archived in figshare or zenodo (and the DOI for the archive)
